feat: enable authentication in routes
